feat(a2a): FR-253 replace type: a2a_call with contrib client
Replace dedicated type: a2a_call node with yamlgraph.contrib.a2a_client
invoked via type: python. Net reduction ~1789 lines.

Created:
- yamlgraph/contrib/a2a_client.py (~230 lines)
- tests/unit/test_a2a_contrib_client.py (29 tests)

Deleted:
- yamlgraph/node_factory/a2a_nodes.py (362 lines)
- yamlgraph/linter/patterns/a2a.py (140 lines)
- tests/unit/test_a2a_call_node.py (1235 lines)

Also implements FR-252: variables: resolution on type: python nodes.
