+++ capabilities/CAP-105-a2a-consumer-phase2.yaml
@@ -1,57 +1,52 @@
 id: CAP-105
 name: A2A Consumer Phase 2 — Agent Card, Skill Selection & Streaming
 description: >
-  Extends the a2a_call node (FR-240) with Agent Card discovery via sync
-  httpx.get() to /.well-known/agent.json, ContextVar-scoped caching per graph
-  invocation, skill selection validated against Agent Card skills at runtime,
-  and SSE streaming via A2AClient.send_message_streaming() in a dedicated
-  thread. Linter W901 warns when skill field is present (runtime-only
-  validation). Linter E904 errors when streaming: true is used on non-a2a_call
-  nodes.
+  A2A consumer features in yamlgraph.contrib.a2a_client: Agent Card discovery
+  via sync httpx.get() to /.well-known/agent.json, ContextVar-scoped caching
+  per graph invocation, skill selection validated against Agent Card skills at
+  runtime, and SSE streaming via A2AClient.send_message_streaming() in a
+  dedicated thread. Replaces dedicated a2a_call node type linter checks
+  (W901/E904) with runtime validation in contrib function (FR-253).
 modules:
-  - yamlgraph/node_factory/a2a_nodes.py
-  - yamlgraph/models/graph_schema.py
-  - yamlgraph/linter/patterns/a2a.py
+  - yamlgraph/contrib/a2a_client.py
 requirements:
   - id: REQ-YG-250
     description: >
-      a2a_call node fetches Agent Card via sync httpx.get() to
-      {agent_url}/.well-known/agent.json; parsed into SDK AgentCard model;
-      cached per agent_url within graph invocation using ContextVar; cache
-      isolated across invocations in long-running processes (MCP server,
-      FastAPI); timeout configurable
+      send_a2a_message() fetches Agent Card via sync httpx.get() to
+      {agent_url}/.well-known/agent.json; parsed into SDK AgentCard model
+      via ParseDict; cached per agent_url within graph invocation using
+      ContextVar; cache isolated across invocations; timeout configurable
     modules:
-      - yamlgraph/node_factory/a2a_nodes.py
-      - tests/unit/test_a2a_call_node.py
+      - yamlgraph/contrib/a2a_client.py
+      - tests/unit/test_a2a_contrib_client.py
   - id: REQ-YG-251
     description: >
-      skill field on a2a_call node selects a specific agent skill; validated
+      skill parameter in state selects a specific agent skill; validated
       against Agent Card skills at runtime; ValueError raised on skill ID miss
-      with available skills listed in error message; skill field added to
-      NodeConfig in graph_schema.py; no card fetch when skill not specified
+      with available skills listed in error message; no card fetch when skill
+      not specified
     modules:
-      - yamlgraph/node_factory/a2a_nodes.py
-      - yamlgraph/models/graph_schema.py
-      - tests/unit/test_a2a_call_node.py
+      - yamlgraph/contrib/a2a_client.py
+      - tests/unit/test_a2a_contrib_client.py
   - id: REQ-YG-252
     description: >
-      streaming: true on a2a_call node uses A2AClient.send_message_streaming()
+      streaming: true in state uses A2AClient.send_message_streaming()
       via dedicated thread with own event loop; requires
-      card.capabilities.streaming == True; result written as complete string to
-      state_key; streaming events logged at DEBUG level; streaming field added
-      to NodeConfig; transport-only (not FR-030 graph-level streaming)
+      card.capabilities.streaming == True; result returned as complete string;
+      streaming events logged at DEBUG level; transport-only (not FR-030
+      graph-level streaming)
     modules:
-      - yamlgraph/node_factory/a2a_nodes.py
-      - yamlgraph/models/graph_schema.py
-      - tests/unit/test_a2a_call_node.py
+      - yamlgraph/contrib/a2a_client.py
+      - tests/unit/test_a2a_contrib_client.py
   - id: REQ-YG-253
     description: >
-      Linter check W901 warns when skill field is present on a2a_call node
-      (informational — runtime-only validation); linter check E904 errors when
-      streaming: true is used on a non-a2a_call node type;
-      check_streaming_on_wrong_type() function; integrated into
-      check_a2a_call_patterns() graph scan
+      Dedicated type: a2a_call node type replaced by type: python +
+      yamlgraph.contrib.a2a_client contrib function; NodeType.A2A_CALL removed
+      from constants; a2a_nodes.py and linter/patterns/a2a.py deleted;
+      W901/E904 linter checks removed (skill/streaming validated at runtime
+      in contrib); FR-252 enables variables: resolution on type: python nodes
     modules:
-      - yamlgraph/linter/patterns/a2a.py
-      - tests/unit/test_a2a_call_node.py
+      - yamlgraph/contrib/a2a_client.py
+      - yamlgraph/tools/python_tool.py
+      - tests/unit/test_a2a_contrib_client.py
 fr: FR-248
