+++ tests/unit/test_node_type_registry.py
@@ -75,7 +75,6 @@ class TestNodeTypeRegistry:
             NodeType.LLM,
             NodeType.ROUTER,
             NodeType.RACE,
-            NodeType.A2A_CALL,
         }
         registered = set(NODE_TYPE_HANDLERS.keys())
         assert expected_types == registered
