+++ yamlgraph/linter/checks.py
@@ -15,7 +15,6 @@ from pydantic import BaseModel
 
 # Valid node types
 VALID_NODE_TYPES = {
-    "a2a_call",
     "agent",
     "copilot",
     "interactive_tool",
