+++ yamlgraph/models/graph_schema.py
@@ -147,26 +147,6 @@ class NodeConfig(BaseModel):
         description="Race candidates: list of {provider, model} dicts",
     )
 
-    # A2A call node fields (FR-240)
-    agent_url: str | None = Field(
-        default=None,
-        description="Base URL of the external A2A agent server",
-    )
-    message: str | None = Field(
-        default=None,
-        description="Jinja2 template for the A2A message text",
-    )
-
-    # A2A call node fields — Phase 2 (FR-248)
-    skill: str | None = Field(
-        default=None,
-        description="Target skill ID on the remote A2A agent",
-    )
-    streaming: bool | None = Field(
-        default=None,
-        description="Use SSE streaming transport for A2A calls (transport-only, not FR-030)",
-    )
-
     # Verification gate (FR-164)
     verification: VerificationConfig | None = Field(
         default=None,
