+++ yamlgraph/tools/python_tool.py
@@ -159,6 +159,7 @@ def create_python_node(
     state_key = node_config.get("state_key", node_name)
     on_error = node_config.get("on_error", "fail")
     loop_limit = node_config.get("loop_limit")
+    variable_templates = node_config.get("variables", {})
 
     # Load the function at node creation time
     func = load_python_function(tool_config)
@@ -178,8 +179,14 @@ def create_python_node(
 
         logger.info(f"🐍 Executing Python node: {node_name} -> {tool_name}")
 
+        # FR-252: Resolve variables expressions before calling function
+        from yamlgraph.utils.expressions import resolve_node_variables
+
+        resolved = resolve_node_variables(variable_templates, state)
+        effective_state = {**state, **resolved} if resolved else state
+
         try:
-            result = func(state)
+            result = func(effective_state)
 
             # If function returns a dict, merge with node metadata
             if isinstance(result, dict):
